put the exec where it belongs
